requirements.txt
celery
channels
channels-redis
daphne
djangorestframework
django-hosts
gunicorn
pyOpenSSL
service-identity